feat: Add rule against punctuations at the end of headings
In document titles and headings, use sentence case. That is, capitalize only
the first word in the title, the first word in a subheading after a colon, and
any proper nouns or other terms that are always capitalized a certain way.

Even though you're using sentence case, don't put a period at the end of a
title or heading.